Fixing ENOENT error
This solution was reported on issue: https://github.com/apla/me.apla.cordova.app-preferences/issues/135 . The solution was implemented in this branch: https://github.com/vash15/me.apla.cordova.app-preferences.
But as we need to have a solution which contains both cordova 9.0.0 compatibility AND a fix for this error. I'm mergin the two solutions into our own repository